bugfix(RasLink-build/common/asterisk-config-check.sh): Keep the extnodes key from growing in rpt.conf
Every time 'system-update' called this script,  the extnodes key was expanded by adding "ext" to the front of the line.
If you wanted to use the AllStarLink node list instead of the PTTLink version, you couldn't.
The line expansion caused the extnodes = key value pair to not match.
This forced your node to use the PTTLink node list.
The extnodes key has been replaced, and will work properly.

Signed-off-by: Jeremy Lincicome [W0JRL] <admin@jlappliedtechnologies.com>